style(MetricWrapper): apply linting and fix style issues
