+++ src/brisk/__init__.py
@@ -10,7 +10,7 @@ from brisk.reporting.report_manager import ReportManager
 from brisk.training.training_manager import TrainingManager
 from brisk.training.workflow import Workflow
 from brisk.utility.algorithm_wrapper import AlgorithmWrapper
-from brisk.utility.MetricWrapper import MetricWrapper
+from brisk.utility.metric_wrapper import MetricWrapper
 from brisk.utility.alert_mailer import AlertMailer
 from brisk.utility.utility import create_metric
 from brisk.utility.logging_util import TqdmLoggingHandler, FileFormatter
