+++ src/brisk/defaults/regression_metrics.py
@@ -11,7 +11,7 @@ import numpy as np
 import scipy
 from sklearn.metrics import _regression
 
-from brisk.utility import MetricWrapper
+from brisk.utility import metric_wrapper
 
 def concordance_correlation_coefficient(
     y_true: np.ndarray,
@@ -39,73 +39,73 @@ def concordance_correlation_coefficient(
 
 
 REGRESSION_METRICS = [
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="explained_variance_score",
         func=_regression.explained_variance_score,
         display_name="Explained Variance Score"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="max_error",
         func=_regression.max_error,
         display_name="Max Error"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="mean_absolute_error",
         func=_regression.mean_absolute_error,
         display_name="Mean Absolute Error",
         abbr="MAE"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="mean_absolute_percentage_error",
         func=_regression.mean_absolute_percentage_error,
         display_name="Mean Absolute Percentage Error",
         abbr="MAPE"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="mean_pinball_loss",
         func=_regression.mean_pinball_loss,
         display_name="Mean Pinball Loss"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="mean_squared_error",
         func=_regression.mean_squared_error,
         display_name="Mean Squared Error",
         abbr="MSE"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="mean_squared_log_error",
         func=_regression.mean_squared_log_error,
         display_name="Mean Squared Log Error"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="median_absolute_error",
         func=_regression.median_absolute_error,
         display_name="Median Absolute Error"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="r2_score",
         func=_regression.r2_score,
         display_name="R2 Score",
         abbr="R2"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="root_mean_squared_error",
         func=_regression.mean_squared_error,
         display_name="Root Mean Squared Error",
         abbr="RMSE"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="root_mean_squared_log_error",
         func=_regression.mean_squared_log_error,
         display_name="Root Mean Squared Log Error"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="concordance_correlation_coefficient",
         func=concordance_correlation_coefficient,
         display_name="Concordance Correlation Coefficient",
         abbr="CCC"
     ),
-    MetricWrapper.MetricWrapper(
+    metric_wrapper.MetricWrapper(
         name="neg_mean_absolute_error",
         func=_regression.mean_absolute_error,
         display_name="Negative Mean Absolute Error",
